+++ CHANGELOG.md
@@ -3,6 +3,34 @@
 All notable changes to this project will be documented in this file.
 Version changes are pinned to SDK releases.
 
+## [Unreleased]
+
+## [1.23.0]
+
+- Add functionality for referral accounts. ([#357](https://github.com/zetamarkets/sdk/pull/357))
+
+### Breaking
+
+- Remove old client.ts and subclient.ts. ([#357](https://github.com/zetamarkets/sdk/pull/357))
+
+## [1.22.1]
+
+- New asset $STRK. ([#361](https://github.com/zetamarkets/sdk/pull/361))
+
+## [1.22.0]
+
+- Deprecate v1 trigger instructions.
+- Add new take trigger order instruction.
+
+## [1.21.4]
+
+- Improve cancelAllMarketOrders for big transactions. ([#356](https://github.com/zetamarkets/sdk/pull/356))
+- Add txSig to return args of edit trigger order functions. ([#356](https://github.com/zetamarkets/sdk/pull/356))
+
+## [1.21.3]
+
+- Bugfix devnet loading assets in Exchange.load(). ([#355](https://github.com/zetamarkets/sdk/pull/354))
+
 ## [1.21.2]
 
 - Small bugfix for DYM in indexToAsset(). ([#354](https://github.com/zetamarkets/sdk/pull/354))